Refactor: Extract CreateVoiceActivity logic into extension files
Moves image handling, markdown formatting, and posting preparation logic from `CreateVoiceActivity` into dedicated extension files (`CreateVoiceImages.kt`, `CreateVoiceMarkdownFormatting.kt`, `CreateVoicePosting.kt`). This significantly reduces the activity's size and improves code organization and maintainability.
